+++ src/providers/sendgrid.rs
@@ -329,7 +329,11 @@ impl SendGridMailer {
     }
 }
 
-#[async_trait]
+#[cfg_attr(
+    all(target_family = "wasm", target_os = "unknown"),
+    async_trait(?Send)
+)]
+#[cfg_attr(not(all(target_family = "wasm", target_os = "unknown")), async_trait)]
 impl Mailer for SendGridMailer {
     fn prepare_email(
         &self,